refactor: move layer script
